updated some of the components and shorthands utils plus composables
