+++ plugins/state.ts
@@ -1,11 +1,11 @@
 export default defineNuxtPlugin(async (nuxtApp) => {
-  const auth = await useAuth()
+  const auth = useAuth()
+  const { tags } = useTag()
   await useUser()
   const cosmeticsStore = useCosmetics()
-  const tagsStore = useTags()
 
   nuxtApp.provide('auth', auth)
   nuxtApp.provide('cosmetics', cosmeticsStore.value)
-  nuxtApp.provide('tag', tagsStore.value)
+  nuxtApp.provide('tag', tags)
   nuxtApp.provide(' notify', notif => addNotification(notif))
 })
